pyproject.toml
[tool.poetry]
name = "chainlit_graphql"
version = "0.1.0"
keywords = ['chainlit private server']
description = "graphql api persist chat data for chainlit."
authors = ["bitacoraai"]
repository = "https://github.com/bitacora-ai"
license = "Apache-2.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlmodel = "^0.0.16"
fastapi = "^0.110.0"
graphql-core = "^3.2.3"
annotated-types = "^0.6.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
strawberry-graphql = "^0.220.0"
tenacity = "^8.2.3"
boto3 = "^1.34.59"
uvicorn = "^0.28.0"
injector = "^0.21.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.2"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.5"
flake8 = "^7.0.0"
black = "^24.2.0"
pytest = "^8.0.2"
asyncio = "^3.4.3"
aiosqlite = "^0.20.0"
pytest-docker = "^3.1.1"
pytest-mock = "^3.12.0"
tree-format = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
